pyproject.toml
[tool.poetry]
name = "bot-service"
version = "0.1.0"
description = ""
authors = ["Squilix <Squilix@no-reply.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.13.1"
loguru = "^0.7.3"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
